+++ packages/studio-base/src/panels/ImageView/storySupport/annotations.ts
@@ -179,7 +179,6 @@ const markers: ImageMarker[] = [
   }),
 ];
 
-// ts-prune-ignore-next
 const annotations = normalizeAnnotations({ markers }, "foxglove_msgs/ImageMarkerArray") ?? [];
 
 export { annotations };
