+++ packages/studio-base/src/panels/ImageView/storySupport/cameraInfo.ts
@@ -1,0 +1,26 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { CameraInfo } from "@foxglove/studio-base/types/Messages";
+
+const cameraInfo: CameraInfo = {
+  width: 400,
+  height: 300,
+  distortion_model: "plumb_bob",
+  D: [-0.437793, 0.183639, -0.003738, -0.001327, 0],
+  K: [2339.067676, 0, 903.297282, 0, 2323.624869, 566.425547, 0, 0, 1],
+  R: [1, 0, 0, 0, 1, 0, 0, 0, 1],
+  P: [2170.145996, 0, 899.453592, 0, 0, 2275.496338, 568.217702, 0, 0, 0, 1, 0],
+  binning_x: 1,
+  binning_y: 1,
+  roi: {
+    x_offset: 0,
+    y_offset: 0,
+    height: 0,
+    width: 0,
+    do_rectify: false,
+  },
+};
+
+export { cameraInfo };
